MeshFlow/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.10
digeo
plotly
matplotlib
scipy
numpy
pyyaml
tqdm
libigl
